+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.4.27] - 2020-04-10
+### Added
+- [ExchangeChannel] TimeFrameExchangeChannel class
+
 ### Fixed
 - [Channel] time frame filter